INS-3465: move bloat tests to a separate tag.
