+++ functest/logicrunner_test.go
@@ -20,7 +20,6 @@ package functest
 
 import (
 	"fmt"
-	"strings"
 	"sync"
 	"testing"
 	"time"
@@ -736,85 +735,6 @@ func (c *TestSagaSelfCallContract) Rollback(delta int) error {
 	require.True(t, checkPassed)
 }
 
-func TestNoWaitCall(t *testing.T) {
-	var contractOneCode = `
-package main
-
-import "github.com/insolar/insolar/logicrunner/builtin/foundation"
-import "github.com/insolar/insolar/insolar"
-import two "github.com/insolar/insolar/application/proxy/basic_notification_call_two"
-
-type One struct {
-	foundation.BaseContract
-	Friend insolar.Reference
-}
-
-func New() (*One, error) {
-	return &One{}, nil
-}
-
-var INSATTR_Hello_API = true
-func (r *One) Hello() error {
-	holder := two.New()
-
-	friend, err := holder.AsChild(r.GetReference())
-	if err != nil {
-		return err
-	}
-
-	r.Friend = friend.GetReference()
-
-	err = friend.MultiplyNoWait()
-	if err != nil {
-		return err
-	}
-
-	return nil
-}
-
-var INSATTR_Value_API = true
-func (r *One) Value() (int, error) {
-	return two.GetObject(r.Friend).GetValue()
-}
-`
-
-	var contractTwoCode = `
-package main
-
-import (
-	"fmt"
-
-	"github.com/insolar/insolar/logicrunner/builtin/foundation"
-)
-
-type Two struct {
-	foundation.BaseContract
-	X int
-}
-
-func New() (*Two, error) {
-	return &Two{X:322}, nil
-}
-
-var INSATTR_Multiply_API = true
-func (r *Two) Multiply() (string, error) {
-	r.X *= 2
-	return fmt.Sprintf("Hello %d times!", r.X), nil
-}
-
-var INSATTR_GetValue_API = true
-func (r *Two) GetValue() (int, error) {
-	return r.X, nil
-}
-`
-	uploadContractOnce(t, "basic_notification_call_two", contractTwoCode)
-	obj := callConstructor(t, uploadContractOnce(t, "basic_notification_call_one", contractOneCode), "New")
-
-	resp := callMethodNoChecks(t, obj, "Hello")
-	require.NotEmpty(t, resp.Error)
-	require.Contains(t, resp.Error.Error(), "reason request is not closed for a detached call")
-}
-
 func TestContextPassing(t *testing.T) {
 	var contractOneCode = `
 package main
@@ -867,35 +787,6 @@ func (r *One) Kill() error {
 
 	resp := callMethod(t, obj, "Kill")
 	require.Empty(t, resp.Error)
-}
-
-// Make sure that panic() in a contract causes a system error and that this error
-// is returned by API.
-func TestPanic(t *testing.T) {
-	var panicContractCode = `
-package main
-
-import "github.com/insolar/insolar/logicrunner/builtin/foundation"
-
-type One struct {
-	foundation.BaseContract
-}
-
-func New() (*One, error) {
-	return &One{}, nil
-}
-
-var INSATTR_Panic_API = true
-func (r *One) Panic() error {
-	panic("AAAAAAAA!")
-	return nil
-}
-`
-	prototype := uploadContractOnce(t, "panic", panicContractCode)
-	obj := callConstructor(t, prototype, "New")
-
-	resp := callMethodNoChecks(t, obj, "Panic")
-	require.Contains(t, resp.Error.Message, "executor error: problem with API call: AAAAAAAA!")
 }
 
 func TestErrorInterface(t *testing.T) {
@@ -1141,50 +1032,6 @@ func New() (*Two, error) {
 	require.Contains(t, resp.Result.Error.Error(), "Epic fail in two.New()")
 }
 
-func TestRecursiveCallError(t *testing.T) {
-	var contractOneCode = `
-package main
-
-import (
-	"github.com/insolar/insolar/logicrunner/builtin/foundation"
-	recursive "github.com/insolar/insolar/application/proxy/recursive_call_one"
-)
-type One struct {
-	foundation.BaseContract
-}
-
-func New() (*One, error) {
-	return &One{}, nil
-}
-
-var INSATTR_Recursive_API = true
-func (r *One) Recursive() (error) {
-	remoteSelf := recursive.GetObject(r.GetReference())
-	err := remoteSelf.Recursive()
-	return err
-}
-
-`
-	protoRef := uploadContractOnce(t, "recursive_call_one", contractOneCode)
-
-	// for now Recursive calls may cause timeouts. Dont remove retries until we make new loop detection algorithm
-	var err string
-	for i := 0; i <= 5; i++ {
-		obj := callConstructor(t, protoRef, "New")
-		resp := callMethodNoChecks(t, obj, "Recursive")
-
-		err = resp.Error.Error()
-		if !strings.Contains(err, "timeout") {
-			// system error is not timeout, loop detected is in response
-			err = resp.Result.ExtractedError
-			break
-		}
-	}
-
-	require.NotEmpty(t, err)
-	require.Contains(t, err, "loop detected")
-}
-
 func TestGetParent(t *testing.T) {
 	var contractOneCode = `
  package main
@@ -1364,79 +1211,6 @@ func (r *Two) GetCounter() (int, error) {
 	resp := callMethodNoChecks(t, obj, "IncrementBy100")
 	require.NotEmpty(t, resp.Error)
 	require.Contains(t, resp.Error.Error(), "reason request is not closed for a detached call")
-}
-
-func TestPrototypeMismatch(t *testing.T) {
-	testContract := `
-package main
-
-import (
-	"github.com/insolar/insolar/logicrunner/builtin/foundation"
-	first "github.com/insolar/insolar/application/proxy/prototype_mismatch_first"
-	"github.com/insolar/insolar/insolar"
-)
-
-func New() (*Contract, error) {
-	return &Contract{}, nil
-}
-
-type Contract struct {
-	foundation.BaseContract
-}
-
-var INSATTR_Test_API = true
-func (c *Contract) Test(firstRef *insolar.Reference) (string, error) {
-	return first.GetObject(*firstRef).GetName()
-}
-`
-
-	// right contract
-	firstContract := `
-package main
-
-import (
-	"github.com/insolar/insolar/logicrunner/builtin/foundation"
-)
-
-type First struct {
-	foundation.BaseContract
-}
-
-var INSATTR_GetName_API = true
-func (c *First) GetName() (string, error) {
-	return "first", nil
-}
-`
-
-	// malicious contract with same method signature and another behaviour
-	secondContract := `
-package main
-
-import (
-	"github.com/insolar/insolar/logicrunner/builtin/foundation"
-)
-
-type First struct {
-	foundation.BaseContract
-}
-
-func New() (*First, error) {
-	return &First{}, nil
-}
-
-var INSATTR_GetName_API = true
-func (c *First) GetName() (string, error) {
-	return "YOU ARE ROBBED!", nil
-}
-`
-
-	uploadContractOnce(t, "prototype_mismatch_first", firstContract)
-	secondObj := callConstructor(t, uploadContractOnce(t, "prototype_mismatch_second", secondContract), "New")
-	testObj := callConstructor(t, uploadContractOnce(t, "prototype_mismatch_test", testContract), "New")
-
-	resp := callMethodNoChecks(t, testObj, "Test", *secondObj)
-	require.Empty(t, resp.Error)
-	require.Contains(t, resp.Result.Error.S, "try to call method of prototype as method of another prototype")
 }
 
 func TestImmutableAnnotation(t *testing.T) {
